Fix failing unit tests
The problem was that the prepared statements were static to take advantage of
thread-local variables to avoid locks. However, the unit tests create many
in-memory databases. Having statements that are static to the class doesn't work
for this. The main application was fine because only one instance of the
BuildState class is ever created.

Statements are now prepared when the class is created. However, since they are
no longer thread-local, we must wrap their usage in synchronized statements.
This hasn't had any noticeable effect on performance.