Add '--delay' option for use with --autopilot
Having a slight delay allows many changed files to be accumulated and then built
in one go. This is useful, for example, when pulling down the latest changes
from Git. The default is 50 milliseconds so that it is long enough to catch
automated changes, but short enough to not irritate the user.