Change project name to Button
The old name (Brilliant Build) never felt right. The name Button has a number of
characteristics that makes it more desireable: it is relatively short,
memorable, and is not used by any major project. It gives a sense of one of the
main goals of this build system - to build with a single command (or at the push
of a *button*). It also has a simple theme that spin-off projects can build on.

Hopefully this one sticks.