Remove resource timestamp comparisons
These are unreliable for detecting changes to directories. The only
correct way to detect changes to directories is to look at the contents.

Thus, comparing the last modified timestamp of files is removed. This
could have a significant performance impact on large projects. In those
cases, a daemon could be implemented to collect changes in the
background as they happen.